rules/string: spell stoi's out-parameter usize, not u64
src.cpp declares `std::size_t *a1`; tgt_unsafe.rs hand-wrote
`*mut u64` for f55/f56/f57.  The two happened to agree only until the
ScalarSugar fix made the call site emit `*mut usize` consistently, at
which point the mismatch surfaced as E0308.  The refcount overlay was
already correct (`Ptr<usize>`).

The three casts writing back through the pointer are now unnecessary and
are dropped rather than retargeted.

Verified end to end, not by inspection: a probe translating std::stoi /
std::stod / std::stoll with a non-null `pos` builds as a cargo crate in
both models and agrees with the same C++ compiled by clang++ on all
seven cases -- including the out-parameter, which is the part that was
broken.  Leading whitespace, a sign, an 0x prefix and a base-16 body all
land the same index one past the last character consumed.

lit unchanged: 931/950 pass, 19 expected-fail, 0 unexpected.
check_rules.py agrees across all three IRs.

Signed off without GPG: signing is unavailable in this worktree.

+++ rules/string/tgt_unsafe.rs
@@ -413,8 +413,8 @@ unsafe fn f54(a0: f64) -> Vec<libc::c_char> {
         .collect::<Vec<libc::c_char>>()
 }
 
-unsafe fn f55(a0: Vec<libc::c_char>, a1: *mut u64, a2: i32) -> i32 {
-    let __pos: *mut u64 = a1;
+unsafe fn f55(a0: Vec<libc::c_char>, a1: *mut usize, a2: i32) -> i32 {
+    let __pos: *mut usize = a1;
     let __text: String = a0
         .iter()
         .copied()
@@ -444,13 +444,13 @@ unsafe fn f55(a0: Vec<libc::c_char>, a1: *mut u64, a2: i32) -> i32 {
     let __digits: String = [&__body[..__sign], &__rest[__skip..__skip + __n]].concat();
     let __value = i64::from_str_radix(&__digits, __radix).expect("stoi: out of range");
     if !__pos.is_null() {
-        *__pos = (__lead + __sign + __skip + __n) as u64;
+        *__pos = __lead + __sign + __skip + __n;
     }
     __value as i32
 }
 
-unsafe fn f56(a0: Vec<libc::c_char>, a1: *mut u64, a2: i32) -> i64 {
-    let __pos: *mut u64 = a1;
+unsafe fn f56(a0: Vec<libc::c_char>, a1: *mut usize, a2: i32) -> i64 {
+    let __pos: *mut usize = a1;
     let __text: String = a0
         .iter()
         .copied()
@@ -480,13 +480,13 @@ unsafe fn f56(a0: Vec<libc::c_char>, a1: *mut u64, a2: i32) -> i64 {
     let __digits: String = [&__body[..__sign], &__rest[__skip..__skip + __n]].concat();
     let __value = i64::from_str_radix(&__digits, __radix).expect("stoll: out of range");
     if !__pos.is_null() {
-        *__pos = (__lead + __sign + __skip + __n) as u64;
+        *__pos = __lead + __sign + __skip + __n;
     }
     __value
 }
 
-unsafe fn f57(a0: Vec<libc::c_char>, a1: *mut u64) -> f64 {
-    let __pos: *mut u64 = a1;
+unsafe fn f57(a0: Vec<libc::c_char>, a1: *mut usize) -> f64 {
+    let __pos: *mut usize = a1;
     let __text: String = a0
         .iter()
         .copied()
@@ -500,7 +500,7 @@ unsafe fn f57(a0: Vec<libc::c_char>, a1: *mut u64) -> f64 {
         .find(|&__i| __body[..__i].parse::<f64>().is_ok())
         .expect("stod: no conversion");
     if !__pos.is_null() {
-        *__pos = (__lead + __n) as u64;
+        *__pos = __lead + __n;
     }
     __body[..__n].parse::<f64>().unwrap()
 }
